feat(contrib/jsternberg): add an influxdb.select function
This adds an `influxdb.select` function as an alternative to using
`from` for pivoted data. At the moment, a person has to do this:

    import "influxdata/influxdb/v1"

    from(bucket: "telegraf")
        |> range(start: -5m)
        |> filter(fn: (r) => r._measurement == "m0" and (r._field == "f0" or r._field == "f1"))
        |> v1.fieldsAsCols()

This replaces this pattern with a specialized function:

    import "contrib/jsternberg/influxdb"

    influxdb.select(from: "telegraf", start: -5m, m: "m0", fields: ["f0", "f1"])

This makes it easier to write the equivalent query. It also helpfully
masks the `_start` and `_stop` columns and effectively removes them from
the group key without incurring the cost of doing a `drop()` call.